% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxreg-package.R
\name{beta_CVX}
\alias{beta_CVX}
\alias{x}
\alias{y}
\title{Simulated data for the cxreg vignette}
\format{
Data objects used to demonstrate features in the cxreg vignette
}
\description{
Simple simulated data, used to demonstrate the features of cxreg
}
\details{
These datasets are artificial, and are used to test out some of the
features of cxreg.
}
\examples{

data(QuickStartExample)
x <- QuickStartExample$x; y <- QuickStartExample$y
classo(x, y)

}
\keyword{datasets}
