% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPremat.default.R
\name{buildPredmat.default}
\alias{buildPredmat.default}
\title{Build Prediction Matrix}
\usage{
\method{buildPredmat}{default}(outlist, lambda, x, foldid, alignment)
}
\description{
Build Prediction Matrix
}
