% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cglassoFlex.R
\name{cglasso.path}
\alias{cglasso.path}
\title{fit a complex-valued graphical Lasso for a path of lambda values}
\usage{
cglasso.path(
  S,
  D,
  type,
  m,
  lambda,
  nlambda,
  lambda.min.ratio,
  W.init,
  stopping_rule,
  stop_criterion,
  maxit,
  thresh,
  trace.it,
  ...
)
}
\arguments{
\item{S}{p x p-dimensional symmetric spectral density (or spectral coherence) matrix. S is considered as being computed by average smoothed periodogram (the bandwidth is computed by using the given nobs).}

\item{D}{The p x p-dimensional diagonal matrix to produce the partial spectral coherence matrix from S. Default is \code{NULL}. If D is not specified, the function automatically takes the inverse square root of the diagonals of S.}

\item{type}{Logical flag to choose the formulation to solve. Default is \code{I}. If type is \code{I}, the algorithm solves CGLASSO-I in the reference,
\deqn{D^{-1/2}\Bigl(\arg\min_{\Theta}\mathrm{Tr}[\hat{R}\hat{\Theta}]
  -\log\det\Theta + \sum_{i\neq j}|\Theta_{ij}|\Bigr)D^{-1/2}}
for the given \eqn{D}. If type is \code{II}, the algorithm solves CGLASSO-II
in the reference: for each iterative classo with covariate update, the scale
matrix \eqn{D} is multiplied. Please see the reference for details.}

\item{m}{Window size. This quantity is need to compute the Fourier frequency, extended BIC, and bandwidth for the average smoothed periodogram.}

\item{lambda}{A user supplied \code{lambda} sequence.
Typical usage is to have the program compute its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}.
Supplying a value of \code{lambda} overrides this.
WARNING: use with care. Avoid supplying a single value for \code{lambda}}

\item{nlambda}{The number of \code{lambda} values - default is 50.}

\item{lambda.min.ratio}{Smallest value for \code{lambda}, as a fraction of
\code{lambda.max}, the (data derived) entry value (i.e. the smallest value
for which all coefficients are zero). ???}

\item{W.init}{Logical flag whether the initially estimated spectral density matrix is given. Default is \code{NULL}.}

\item{stopping_rule}{Logical flag if the algorithm is terminated by stopping rule. If the algorithm is early terminated, not all estimates for initially designated lambdas are explored.}

\item{stop_criterion}{Stopping criterion for early termination. Default is \code{EBIC} (Extended BIC). Alternatively, \code{AIC} (AIC) and \code{RMSE} (root mean squared error between two consecutive estimates) can be used.}

\item{maxit}{Maximum number of iterations of both outer and inner loops. Default 500.}

\item{thresh}{Convergence threshold for coordinate descent. Default is 1e-4.}

\item{trace.it}{If \code{trace.it=1}, then a progress bar is displayed;
useful for big models that take a long time to fit.}

\item{\dots}{Other arguments that can be passed to \code{cglasso}}
}
\value{
An object with class \code{"cglassofit"} and \code{"cglasso"}.
\describe{
\item{stop_arr}{Sequence of values of the information criterion for each lambda.}
\item{stop_criterion}{Stopping criterion used.}
\item{min_index}{The index for the lambda that minimizes the information criterion.}
\item{lambda_grid}{Sequence of lambdas used.}
\item{Theta_list}{Estimated spectral precision matrix for each fixed lambda, provided as a list.}
\item{type}{Type of the formulation used, either CGLASSO-I or CGLASSO-II.}
\item{D}{Scale diagonal matrix used.}
}
}
\description{
Fit a complex-valued graphical Lasso formulation for a path of lambda values.
\code{cglasso.path} solves the penalized Gaussian maximum likelihood problem for a path of lambda values.
}
