% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxreg-package.R
\name{classo-internal}
\alias{classo-internal}
\alias{cvtype}
\alias{cvstats}
\title{Internal classo functions}
\description{
These are not intended for use by users. \code{lambda.interp} does linear
interpolation of the lambdas to obtain a prediction at a new point s.
\code{nonzeroCoef} determines in an efficient manner which variables are
nonzero in each fit.
}
\author{
Younghoon Kim
}
\keyword{internal}
