% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classoFlex.R
\name{classo.fit}
\alias{classo.fit}
\title{Fit a classo for a single value of lambda}
\usage{
classo.fit(
  x,
  y,
  weights,
  lambda,
  family = gaussian(),
  intercept = FALSE,
  thresh = 1e-10,
  maxit = 1e+05,
  warm = NULL,
  from.classo.path = FALSE,
  save.fit = FALSE,
  trace.it = 0
)
}
\arguments{
\item{x}{Input matrix, of dimension \code{nobs x nvars}; each row is an
observation vector. If it is a sparse matrix, it is assumed to be unstandardized.
It should have attributes \code{xm} and \code{xs}, where \code{xm(j)} and
\code{xs(j)} are the centering and scaling factors for variable j respectively.
If it is not a sparse matrix, it is assumed that any standardization needed
has already been done.}

\item{y}{Quantitative response variable.}

\item{weights}{Observation weights. \code{classo.fit} does NOT standardize
these weights.}

\item{lambda}{A single value for the \code{lambda} hyperparameter.}

\item{intercept}{Should intercept set to zero (default=FALSE) or be fitted (TRUE)?}

\item{thresh}{Convergence threshold for coordinate descent. Each inner
coordinate-descent loop continues until the maximum change in the objective
after any coefficient update is less than thresh times the null deviance.
Default value is \code{1e-10}.}

\item{maxit}{Maximum number of passes over the data; default is \code{10^5}.
(If a warm start object is provided, the number of passes the warm start object
performed is included.)}

\item{warm}{Either a \code{classofit} object or a list (with names \code{beta}
and \code{a0} containing coefficients and intercept respectively) which can
be used as a warm start. Default is \code{NULL}, indicating no warm start.
For internal use only.}

\item{from.classo.path}{Was \code{classo.fit()} called from \code{classo.path()}?
Default is FALSE.}

\item{save.fit}{Return the warm start object? Default is FALSE.}

\item{trace.it}{Controls how much information is printed to screen. If
\code{trace.it=2}, some information about the fitting procedure is printed to
the console as the model is being fitted. Default is \code{trace.it=0}
(no information printed). (\code{trace.it=1} not used for compatibility with
\code{classo.path}.)}
}
\value{
An object with class "classofit" and "classo". The list
returned contains more keys than that of a "classo" object.
\item{a0}{Intercept value.}
\item{beta}{A \code{nvars x 1} matrix of coefficients, stored in sparse matrix
format.}
\item{df}{The number of nonzero coefficients.}
\item{dim}{Dimension of coefficient matrix.}
\item{lambda}{Lambda value used.}
\item{dev.ratio}{The fraction of (null) deviance explained. The deviance
calculations incorporate weights if present in the model. The deviance is
defined to be 2*(loglike_sat - loglike), where loglike_sat is the log-likelihood
for the saturated model (a model with a free parameter per observation).
Hence dev.ratio=1-dev/nulldev.}
\item{nulldev}{Null deviance (per observation). This is defined to be
2*(loglike_sat -loglike(Null)). The null model refers to the intercept model.}
\item{npasses}{Total passes over the data.}
\item{jerr}{Error flag, for warnings and errors (largely for internal
debugging).}
\item{call}{The call that produced this object.}
\item{nobs}{Number of observations.}
\item{warm_fit}{If \code{save.fit=TRUE}, output of C++ routine, used for
warm starts. For internal use only.}
\item{family}{Family used for the model.}
\item{converged}{A logical variable: was the algorithm judged to have
converged?}
\item{obj_function}{Objective function value at the solution.}
}
\description{
WARNING: Users should not call \code{classo.fit} directly.
Higher-level functions in this package call \code{classo.fit} as a subroutine.
If a warm start object is provided, some of the other arguments in the function may be overriden.
}
\details{
\code{classo.fit} solves the complex-valued lasso  problem for a single, user-specified
value of lambda.

In terms of standardization: \code{classo.fit} does not standardize \code{x}
and \code{weights}.
}
