% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classo_cov_fixed_YK.R
\name{classo_cov_fixed}
\alias{classo_cov_fixed}
\title{classo_cov_fixed}
\usage{
classo_cov_fixed(X, Y, lambda, weights, b.init = NULL)
}
\arguments{
\item{X}{input1}

\item{Y}{input2}

\item{lambda}{input3}

\item{weights}{input4}

\item{b.init}{input5}
}
\value{
output
}
\description{
cglasso with covariate update
}
\examples{
example <- classo_cov_fixed(X, Y, lambda, weights)

}
