% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv.classo.R
\name{coef.cv.classo}
\alias{coef.cv.classo}
\title{Extract coefficients from a cv.classo object}
\usage{
\method{coef}{cv.classo}(object, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{Fitted \code{"cv.classo"} object.}

\item{s}{Value(s) of the penalty parameter \code{lambda}. Default is
\code{"lambda.1se"}. Alternatively \code{"lambda.min"} can be used,
or a numeric value (or vector) of \code{lambda} directly.}

\item{\dots}{Additional arguments passed to \code{\link{coef.classo}}.}
}
\value{
A complex-valued coefficient matrix; see \code{\link{predict.classo}}.
}
\description{
A convenience wrapper that extracts coefficients from the
\code{classo.fit} stored in a \code{"cv.classo"} object at a chosen
lambda value.
}
\seealso{
\code{\link{classo}}, \code{\link{coef.classo}},
\code{\link{cv.classo}}.
}
\keyword{models}
\keyword{regression}
