% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxreg-package.R
\name{cxreg}
\alias{cxreg}
\alias{x}
\alias{y}
\title{Simulated data for the cxreg vignette}
\format{
Data objects used to demonstrate features in the cxreg vignette
}
\description{
Simple simulated data, used to demonstrate the features of cxreg
}
\details{
These datasets are artificial, and are used to test out some of the
features of cxreg.
}
\examples{
\donttest{ 
data(classo_example)
x <- classo_example$x
y <- classo_example$y
classo(x,y)

data(cglasso_example)
f_hat <- cglasso_example$f_hat
n <- cglasso_example$n
cglasso(S=f_hat,type="I",nobs=n)
cglasso(S=f_hat,type="II",nobs=n)
}
}
\keyword{datasets}
