% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.functions.R
\name{dft.all}
\alias{dft.all}
\title{Discrete Fourier Transform of matrix X}
\usage{
dft.all(X)
}
\arguments{
\item{X}{A numeric matrix of size \eqn{nobs \times nvar}, where DFT is applied across the rows (time points).}
}
\value{
A complex-valued matrix of dimension \eqn{nobs \times nvar} representing all DFT components of the original matrix.
}
\description{
Computes the (normalized) discrete Fourier transform (DFT) of a matrix \code{X} row-wise using \code{mvfft} over all Fourier frequencies.
}
