% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.functions.R
\name{fhat_at}
\alias{fhat_at}
\title{Smoothed periodogram at a fixed frequency}
\usage{
fhat_at(dft, j, m)
}
\arguments{
\item{dft}{A complex-valued matrix of dimension \eqn{nobs \times nvar},
typically the output of \code{\link{dft.all}}.}

\item{j}{An integer index in \eqn{1,\ldots,nobs} at which the smoothed
periodogram is evaluated.}

\item{m}{A non-negative integer specifying the window half-width. Frequency
indices outside \eqn{[1, nobs]} are wrapped circularly.}
}
\value{
A \eqn{nvar \times nvar} complex-valued Hermitian matrix giving the
smoothed periodogram estimate at frequency \eqn{\omega_j}.
}
\description{
Extracts a window of \code{2m + 1} DFT components centered at index \code{j}
from a pre-computed full DFT matrix and returns their averaged outer product,
i.e. the smoothed periodogram (spectral density) estimate at frequency
\eqn{\omega_j = 2\pi j / n}:
\deqn{\hat{f}(\omega_j) = \frac{1}{2m+1} \sum_{k=j-m}^{j+m} d_k d_k^*,}
where \eqn{d_k} is the \eqn{k}-th row of the DFT matrix (as a column vector)
and \eqn{*} denotes the conjugate transpose.
}
\examples{
p <- 30
n <- 500
set.seed(1010)
X_t <- matrix(rnorm(n * p), n, p)
dft  <- dft.all(X_t)
m    <- floor(sqrt(n)); j <- 1
fhat <- fhat_at(dft, j, m)
}
\references{
Deb, N., Kim Y., Basu, S. (2026)
\emph{Inference for High-Dimensional Sparse Spectral Precision Matrices},
\url{https://arxiv.org/abs/2606.07986}.
}
\seealso{
\code{\link{dft.all}}, \code{\link{dft.j}}
}
\author{
Navonil Deb, Younghoon Kim, Sumanta Basu \cr Maintainer: Younghoon Kim
\email{ykim124@ua.edu}
}
\keyword{density}
\keyword{models}
\keyword{spectral}
