% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complex_operators_YK.R
\name{gen_soft_thrs}
\alias{gen_soft_thrs}
\title{soft_thrs_single}
\usage{
gen_soft_thrs(z, thrs)
}
\arguments{
\item{z}{input1}

\item{thrs}{input2}
}
\value{
output
}
\description{
Generalized soft threshold operator
}
\examples{
example <- gen_soft_thrs(z,thrs)

}
