% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classoFlex.R
\name{get_eta}
\alias{get_eta}
\title{Helper function to get etas (linear predictions)}
\usage{
get_eta(x, beta)
}
\arguments{
\item{x}{Input matrix, of dimension \code{nobs x nvars}; each row is an
observation vector. If it is a sparse matrix, it is assumed to be unstandardized.
It should have attributes \code{xm} and \code{xs}, where \code{xm(j)} and
\code{xs(j)} are the centering and scaling factors for variable j respsectively.
If it is not a sparse matrix, it is assumed to be standardized.}

\item{beta}{Feature coefficients.}
}
\description{
Given x, coefficients and intercept, return linear predictions. Wrapper that
works with both regular and sparse x. Only works for single set of coefficients
and intercept.
}
