% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classoFlex.R
\name{get_start}
\alias{get_start}
\title{Get null deviance, starting mu and lambda max}
\usage{
get_start(x, y, weights, family = gaussian, intercept)
}
\arguments{
\item{x}{Input matrix, of dimension \code{nobs x nvars}; each row is an
observation vector.
It should have attributes \code{xm} and \code{xs}, where \code{xm(j)} and
\code{xs(j)} are the centering and scaling factors for variable j respsectively.
It is assumed to be standardized.}

\item{y}{Quantitative response variable.}

\item{weights}{Observation weights.}

\item{family}{To match the convention in \code{glmnet}, "gaussian" is assumed.
(See \code{\link[stats:family]{family}} for details on family functions.)}

\item{intercept}{Does the model we are fitting have an intercept term or not?}
}
\description{
Return the null deviance, starting mu and lambda max values for
initialization. For internal use only.
}
\details{
This function is called by \code{classo.path} for null deviance,
starting mu and lambda max values.
It is also called by \code{classo.fit} when used without warmstart,
but they only use the null deviance and starting mu values.

When \code{x} is not sparse, it is expected to already by centered and scaled.
When \code{x} is sparse, the function will get its attributes \code{xm} and
\code{xs} for its centering and scaling factors.

Note that whether \code{x} is centered & scaled or not,
the values of \code{mu} and \code{nulldev} don't change.
However, the value of \code{lambda_max} does change,
and we need \code{xm} and \code{xs} to get the correct value.
}
