% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nodewise_YK.R
\name{nodewise_YK}
\alias{nodewise_YK}
\title{classo_warm}
\usage{
nodewise_YK(
  Xt,
  j,
  method = "or",
  type = "ols",
  scale = TRUE,
  weights = NULL,
  pathwise = FALSE
)
}
\arguments{
\item{Xt}{input1}

\item{j}{input2}

\item{method}{input3}

\item{type}{input4}

\item{scale}{input5}

\item{weights}{input6}

\item{pathwise}{input7}
}
\value{
output
}
\description{
Assume columns of X have norm = sqrt(N)
}
\examples{
example <- nodewise_YK(Xt, j, method = "or", type = "ols", scale = TRUE, weights=NULL, pathwise=FALSE)

}
