% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classoFlex.R
\name{pen_function}
\alias{pen_function}
\title{classo penalty value}
\usage{
pen_function(coefficients)
}
\arguments{
\item{coefficients}{The model's coefficients (excluding intercept).}
}
\description{
Returns the classo penalty value.
}
\details{
The penalty is defined as
\deqn{\sum |\beta|.}
Note the omission of the multiplicative \code{lambda} factor.
}
