% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cglasso.R
\name{plot.cglasso}
\alias{plot.cglasso}
\title{Plot heatmap from a "cglasso" object}
\usage{
\method{plot}{cglasso}(
  x,
  index = 1,
  type = c("mod", "real", "imaginary", "both"),
  label = FALSE,
  ...
)
}
\arguments{
\item{x}{Fitted \code{"cglasso"} model, i.e. an object of class
\code{"cglassofit"} returned by \code{\link{cglasso}}.}

\item{index}{Integer index selecting which element of \code{x$Theta_list}
to plot (i.e. which lambda value). Must be in
\eqn{1, \ldots,} \code{length(x$Theta\_list)}. Default is \code{1}.}

\item{type}{Which component to plot: \code{"real"} for the real part,
\code{"imaginary"} for the imaginary part, \code{"mod"} for the modulus
(absolute value), or \code{"both"} for real and imaginary side by side.
Default is \code{"mod"}.}

\item{label}{If \code{TRUE}, label the axes with variable names.}

\item{\dots}{Other graphical parameters passed to \code{image.plot}.}
}
\value{
No return value; called for its side effect of producing a plot.
}
\description{
Produces a heatmap of the estimated spectral precision matrix for a fitted
\code{"cglasso"} object. An inverse spectral matrix heatmap is produced.
}
\seealso{
\code{\link{cglasso}}
}
\author{
Navonil Deb, Younghoon Kim, Sumanta Basu \cr Maintainer: Younghoon Kim
\email{ykim124@ua.edu}
}
