% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.classo.R
\name{plot.classo}
\alias{plot.classo}
\title{plot coefficients from a "classo" object}
\usage{
\method{plot}{classo}(x, xvar = c("norm", "lambda", "dev"), label = FALSE, ...)
}
\arguments{
\item{x}{fitted \code{"classo"} model}

\item{xvar}{What is on the X-axis. \code{"norm"} plots against the L1-norm
of the coefficients, \code{"lambda"} against the log-lambda sequence, and
\code{"dev"} against the percent deviance explained.}

\item{label}{If \code{TRUE}, label the curves with variable sequence
numbers.}

\item{\dots}{Other graphical parameters to plot}
}
\value{
No return value, called for side effects (produces a plot).
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{"classo"} object.
}
\details{
A coefficient profile plot is produced.
}
\seealso{
\code{classo}
}
\author{
Navonil Deb, Younghoon Kim, Sumanta Basu \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
