% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.classo.R
\name{predict.cv.classo}
\alias{predict.cv.classo}
\title{Make predictions from a "cv.classo" object}
\usage{
\method{predict}{cv.classo}(object, newx, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{Fitted \code{"cv.classo"} object.}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be
made. Must be a complex-valued matrix.}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is \code{"lambda.1se"}. Alternatively
\code{"lambda.min"} can be used, or a numeric value (or vector) of
\code{lambda} directly.}

\item{\dots}{Additional arguments passed to the \code{predict} method for
\code{"classo"} objects.}
}
\value{
The object returned by \code{\link{predict.classo}} for the chosen
\code{s}: a complex-valued matrix of predicted values, or a coefficient
matrix, depending on \code{type}.
}
\description{
This function makes predictions from a cross-validated classo model, using
the stored \code{"classo.fit"} object and a chosen lambda value.
}
\seealso{
\code{\link{classo}}, \code{\link{predict.classo}},
\code{\link{coef.cv.classo}}, \code{\link{cv.classo}}.
}
\author{
Younghoon Kim, Navonil Deb, Sumanta Basu \cr Maintainer:
Younghoon Kim \href{mailto:ykim124@ua.edu}{ykim124@ua.edu}
}
\keyword{models}
\keyword{regression}
