% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.classo.R
\name{print.cv.classo}
\alias{print.cv.classo}
\title{print a cross-validated classo object}
\usage{
\method{print}{cv.classo}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted 'cv.classo' object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
The matrix above is silently returned
}
\description{
Print a summary of the results of cross-validation for a classo model.
}
\details{
A two-row table is printed showing the optimal \code{lambda} values
selected by the minimum CV error (\code{lambda.min}) and the 1-standard-error
rule (\code{lambda.1se}), along with the corresponding CV error, its standard
error, and the number of nonzero coefficients.
}
\seealso{
\code{classo}, \code{predict} and \code{coef} methods.
}
\keyword{models}
\keyword{regression}
