% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complex_operators_YK.R
\name{soft_thrs_single}
\alias{soft_thrs_single}
\title{soft_thrs_single}
\usage{
soft_thrs_single(z, thrs)
}
\arguments{
\item{z}{input1}

\item{thrs}{input2}
}
\value{
output
}
\description{
Soft threshold operator (typical lasso operator for a single entry)
}
\examples{
example <- soft_thrs_single(z,thrs)

}
