% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPremat.default.R
\name{buildPredmat.default}
\alias{buildPredmat.default}
\title{Build Prediction Matrix (Default Method)}
\usage{
\method{buildPredmat}{default}(outlist, lambda, x, foldid, alignment, ...)
}
\arguments{
\item{outlist}{A list of fitted models.}

\item{lambda}{A vector of penalty values.}

\item{x}{A predictor matrix.}

\item{foldid}{Fold identifiers.}

\item{alignment}{Alignment method (not used).}

\item{\dots}{Other arguments}
}
\value{
A numeric matrix of predicted values with dimensions
\code{nrow(x)} by \code{length(lambda)}. Rows correspond to
observations and columns correspond to penalty values.
}
\description{
These are not intended for use by users. Constructs a prediction matrix for cross-validation folds, using the coefficient paths in \code{outlist}. Inspired by internal functions in the \pkg{glmnet} package.
}
