% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPremat.default.R
\name{buildPredmat}
\alias{buildPredmat}
\title{Build Prediction Matrix}
\usage{
buildPredmat(outlist, lambda, x, foldid, alignment, ...)
}
\arguments{
\item{outlist}{A list of fitted models.}

\item{lambda}{A vector of penalty values.}

\item{x}{A predictor matrix.}

\item{foldid}{Fold identifiers.}

\item{alignment}{Alignment method (not used).}

\item{\dots}{Other arguments}
}
\value{
The return value depends on the method; for the default method,
see \code{\link{buildPredmat.default}}.
}
\description{
These are not intended for use by users. Constructs a prediction matrix for cross-validation folds, using the coefficient paths in \code{outlist}.
Inspired by internal functions in the \code{glmnet} package.
}
