% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{cglasso_example}
\alias{cglasso_example}
\title{Example Data for CGLASSO}
\format{
A list with components:
\describe{
\item{S}{Spectral density matrix}
\item{n}{Sample size}
}
}
\usage{
data(cglasso_example)
}
\description{
A synthetic dataset used to illustrate complex-valued graphical Lasso.
}
\keyword{datasets}
