% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxreg-package.R
\name{classo-internal}
\alias{classo-internal}
\alias{cvtype}
\alias{cvstats#'}
\title{Internal classo functions}
\description{
These are not intended for use by users. \code{lambda.interp} does linear
interpolation of the lambdas to obtain a prediction at a new point s.
\code{nonzeroCoef} determines in an efficient manner which variables are
nonzero in each fit.
\code{jerr} (not currently available) prints out error messages from the C++ routines.
\code{plotCoef} is called by the \code{plot} method for \code{cxreg}
objects. \code{check_dots} is used in \code{coef} and \code{predict} with
argument \code{exact=TRUE}, to make sure user supplies original data used to
fit the \code{"classo"} object.
}
\author{
Younghoon Kim
}
\keyword{internal}
