% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classo.R
\name{classo}
\alias{classo}
\title{fit a complex-valued lasso}
\usage{
classo(
  x,
  y,
  weights = NULL,
  family = gaussian(),
  lambda = NULL,
  nlambda = 100,
  lambda.min.ratio = ifelse(nobs < nvars, 0.01, 1e-04),
  standardized = FALSE,
  intercept = FALSE,
  maxit = 1e+05,
  trace.it = 0,
  ...
)
}
\arguments{
\item{x}{input matrix, of dimension nobs x nvars; each row is an observation vector.
Requirement: \code{nvars >1}; in other words, \code{x} should have 2 or more columns.}

\item{y}{response variable.}

\item{weights}{observation weights. Default is 1 for each observation.}

\item{family}{To match the conventional functions used in \code{glmnet}, "gaussian" is assumed. This has to be eliminated.}

\item{lambda}{A user supplied \code{lambda} sequence.
Typical usage is to have the program compute its own \code{lambda} sequence based on
\code{nlambda} and \code{lambda.min.ratio}.
Supplying a value of \code{lambda} overrides this.
WARNING: use with care. Avoid supplying a single value for \code{lambda}
(for predictions after CV use \code{predict()} instead).
Supply instead a decreasing sequence of \code{lambda} values.
\code{classo} relies on its warms starts for speed, and its often faster to fit a whole path than compute a single fit.}

\item{nlambda}{The number of \code{lambda} values - default is 100.}

\item{lambda.min.ratio}{Smallest value for \code{lambda}, as a fraction of
\code{lambda.max}, the (data derived) entry value (i.e. the smallest value
for which all coefficients are zero). The default depends on the sample size
\code{nobs} relative to the number of variables \code{nvars}.
If \code{nobs > nvars}, the default is \code{0.0001}, close to zero.
If \code{nobs < nvars}, the default is \code{0.01}.
A very small value of \code{lambda.min.ratio} will lead to a saturated fit
in the \code{nobs < nvars} case.}

\item{intercept}{Should intercept(s) set to zero (default=FALSE) or be fitted (TRUE).}

\item{trace.it}{If \code{trace.it=1}, then a progress bar is displayed;
useful for big models that take a long time to fit.}

\item{standardize}{Logical flag for x variable standardization, prior to
fitting the model sequence. The coefficients are always returned on the
original scale. Default is \code{standardize=FALSE}.}

\item{thresh}{Convergence threshold for coordinate descent.
Each inner coordinate-descent loop continues until the maximum change in the objective
after any coefficient update is less than \code{thresh} times the null
deviance. Defaults value is \code{1E-7}.}
}
\description{
Fit a complex-valued lasso formulation via complex update coordinate descent algorithm.
By defining a field isomophism between complex values and its 2 by 2 representation,
it enables to update each coordinate of the solution as a regular coordinate descent algorithm.
}
\details{
The sequence of models implied by \code{lambda} is fit by coordinate descent.
For \code{family="gaussian"} this is the lasso sequence.
}
\examples{
set.seed(1010)
n = 1000
p = 200
x = array(rnorm(n*p), c(n,p)) + (1+1i) * array(rnorm(n*p), c(n,p))
for (j in 1:p) x[,j] = x[,j] / sqrt(mean(Mod(x[,j])^2))
e = rnorm(n) + (1+1i) * rnorm(n)
b = c(1, -1, rep(0, p-2)) + (1+1i) * c(-0.5, 2, rep(0, p-2))
y = x \%*\% b + e
fit.test = classo(x, y)

}
\references{
Deb, N., Kuceyeski, A., Basu, S. (2024)
\emph{Regularized Estimation of Sparse Spectral Precision Matrices},
\url{https://arxiv.org/abs/2401.11128}.
}
\author{
Navonil Deb, Younghoon Kim, Sumanta Basu \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
\keyword{complex-valued}
\keyword{models}
\keyword{regression}
