% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{classo_example}
\alias{classo_example}
\title{Example Data for CLASSO}
\format{
A list with components:
\describe{
\item{X}{Design matrix}
\item{Y}{Response vector}
}
}
\usage{
data(classo_example)
}
\description{
A toy example demonstrating complex-valued lasso.
}
\keyword{datasets}
