% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxreg-package.R
\name{cxreg-package}
\alias{cxreg-package}
\title{Complex-valued Lasso model paths}
\description{
This package fits complex-valued Lassofor regression using coordinate descent. The algorithm is extremely fast, and exploits sparsity in the input x matrix where it exists.
A variety of predictions can be made from the fitted models.
}
\details{
\tabular{ll}{ Package: \tab cxreg \cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2024-12-01 \cr License: \tab What license is it under?\cr }
Very simple to use. Accepts \code{x,y} data for regression models, and
produces the regularization path over a grid of values for the tuning
parameter \code{lambda}. Only 5 functions: \code{classo}\cr
\code{predict.classo}\cr \code{plot.classo}\cr
\code{coef.classo}
}
\examples{

x <- array(rnorm(100*20), c(100,20)) + (1+1i) * array(rnorm(100*20), c(100,20))
for (j in 1:20) x[,j] <- x[,j] / sqrt(mean(Mod(x[,j])^2))
e <- rnorm(100) + (1+1i) * rnorm(100)
b <- c(1, -1, rep(0, 18)) + (1+1i) * c(-0.5, 2, rep(0, 18))
y <- x \%*\% b + e
fit <- glmnet(x, y)
predict(fit, newx = x[1:5, ], s = c(0.01, 0.005))
predict(fit, type = "coef")
plot(fit, xvar = "lambda")

}
\references{
Deb, N., Kuceyeski, A., Basu, S. (2024)
\emph{Regularized Estimation of Sparse Spectral Precision Matrices},
\url{https://arxiv.org/abs/2401}
}
\author{
Younghoon Kim, Navonil Deb, Sumanta Basu \cr Maintainer:
Younghoon Kim \href{mailto:yk748@cornell.edu}{yk748@cornell.edu}
}
\keyword{models}
\keyword{package}
\keyword{regression}
