% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_misc_YK.R
\name{dft.X}
\alias{dft.X}
\title{dft.X}
\usage{
dft.X(X, j, m)
}
\arguments{
\item{X}{input1}

\item{j}{input2}

\item{m}{input3}
}
\value{
output
}
\description{
discrete Fourier transformation
}
\examples{
example <- dft.X(X, j, m)

}
