% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cglasso.R
\name{plot.cglasso}
\alias{plot.cglasso}
\title{plot heatmap from a "cglasso" object}
\usage{
\method{plot}{cglasso}(
  x,
  index,
  type = c("real", "imaginary", "mod", "both"),
  label = FALSE,
  ...
)
}
\arguments{
\item{x}{fitted \code{"cglasso"} model}

\item{index}{For which inverse spectral matrix profile is the plot to be drawn? Default is 1.
The index must be provided within the length of the sequence of lambdas.}

\item{type}{Whether the plot is for real or imaginary part, or both, or in modulus (mod; absolute scale). Default is \code{mod}.}

\item{label}{If \code{TRUE}, label the axes with variable names.}

\item{\dots}{Other graphical parameters to plot.}
}
\value{
No return value, called for side effects (produces a plot).
}
\description{
Produces plot of the estimated inverse spectral matrix for a fitted
\code{"cglasso"} object. A inverse spectral matrix profile plot is produced.
}
\seealso{
\code{cglasso}
}
\author{
Navonil Deb, Younghoon Kim, Sumanta Basu \cr Maintainer: Younghoon Kim
\email{yk748@cornell.edu}
}
