% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv.classo.R
\name{print.cv.classo}
\alias{print.cv.classo}
\title{print a cross-validated classo object}
\usage{
\method{print}{cv.classo}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted 'cv.classo' object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
The matrix above is silently returned
}
\description{
Print a summary of the results of cross-validation for a classo model.
}
\details{
A summary of the cross-validated fit is produced, slightly different for a
'cv.relaxed' object than for a 'cv.classo' object.  Note that a 'cv.relaxed'
object inherits from class 'cv.classo', so by directly invoking
\code{print.cv.classo(object)} will print the summary as if
\code{relax=TRUE} had not been used.
}
\seealso{
\code{classo}, \code{predict} and \code{coef} methods.
}
\keyword{models}
\keyword{regression}
